Make inbox and unread the main search of interest
This makes it easier to keep track of emails in the inbox without having
to check two searches regularly. It also makes it easier to leave emails
in the inbox, rather than automatically moving them out and then
manually having to move them back, if they require more attention later.